requirements.txt
-f /usr/share/pip-wheels
asn1crypto==0.23.0
Babel==2.5.1
beautifulsoup4==4.6.0
cachetools==2.0.1
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.3
dhl-shipping==2.0.4
Django==1.11.4
django-compressor==2.2	
django-sslify==0.2.7
docx-mailmerge==0.3.0
easypost==3.6.2
enum34==1.1.6
et-xmlfile==1.0.1
exchangelib==1.10.7
facebookads==2.10.1
html5lib==0.999999999
idna==2.6
InvoiceGenerator==0.5.4
ipaddress==1.0.18
jdcal==1.3
Klarna-API==3.1.0
KlarnaCheckout==3.0.1
lxml==4.1.0
mechanize==0.3.6
mysql-connector==2.1.6
MySQL-python==1.2.5
olefile==0.44
openpyxl==2.4.8
paymill==0.1.1
paymill-jsonobject==0.7.1b0
paymill-wrapper==2.3.0
phonenumbers==8.8.0
Pillow==4.3.0
pycparser==2.18
pyOpenSSL==17.3.0
PyPDF2==1.26.0
python-dateutil==2.6.1
pytz==2017.2
qrcode==5.3
qrplatba==0.3.4
reportlab==3.4.0
requests==2.18.4
shippo==1.5.1
sitemap-generator==0.3.9
six==1.11.0
urllib3==1.22
webencodings==0.5.1
XlsxWriter==1.0.0
xmltodict==0.11.0
zenpy==1.2.3
